1221 페이징처리,기관정보

+++ src/main/java/kr/co/kfs/assetedu/repository/Sys01UserRepository.java
@@ -8,7 +8,7 @@ import kr.co.kfs.assetedu.model.Sys01User;
 public interface Sys01UserRepository {
 	// list and totalcount
 	List<Sys01User> selectList(QueryAttr queryAttr);
-	Integer selectCount(QueryAttr queryAttr);
+	Long selectCount(QueryAttr queryAttr);
 	
 	String getUserId();
